Pushed for Visual Studio 2019
Made sure to update compatibility with the newer Visual Studio 2019. But there are errors:

Cannot convert lambda expression to type 'Delegate' because it is not a delegate type.

Field 'BspOptionsForm.components/CreateModForm.components/LightOptionsForm.components' is never assigned to, and will always have its default value null

We will upload backups in case.